+++ doom3/neo/idlib/hashing/MD4.cpp
@@ -38,7 +38,8 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+//k 64
+typedef unsigned int UINT4;
 
 /* MD4 context. */
 typedef struct {
@@ -250,10 +251,11 @@ void MD4_Final(MD4_CTX *context, unsigned char digest[16])
 MD4_BlockChecksum
 ===============
 */
-unsigned long MD4_BlockChecksum(const void *data, int length)
+//k 64
+unsigned int MD4_BlockChecksum(const void *data, int length)
 {
-	unsigned long	digest[4];
-	unsigned long	val;
+	unsigned int	digest[4];
+	unsigned int	val;
 	MD4_CTX			ctx;
 
 	MD4_Init(&ctx);
