+++ doom3/neo/idlib/hashing/MD4.h
@@ -10,6 +10,7 @@
 ===============================================================================
 */
 
-unsigned long MD4_BlockChecksum(const void *data, int length);
+//k 64
+unsigned int MD4_BlockChecksum(const void *data, int length);
 
 #endif /* !__MD4_H__ */
